requirements.txt
flask
flask-cors
tensorflow
numpy
scipy
mne
pillow
python-dotenv
opencv-python
matplotlib
scikit-learn
pymongo
flask-jwt-extended
bcrypt
weasyprint
pdfkit
